+++ packages/marketplace-registry/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/marketplace-registry
 
+## 1.5.3
+
+### Patch Changes
+
+- @inlang/marketplace-manifest@1.3.2
+
 ## 1.5.2
 
 ### Patch Changes